requirements.txt
PyQt5>=5.15
speechrecognition>=3.10.0
pyttsx3>=2.90
google-generativeai>=0.3.2
pyaudio>=0.2.14
comtypes>=1.1.14
pywin32>=305
